workers/pc-worker/requirements.txt
websockets
python-dotenv
psutil
